Tests/LibWeb/Text/expected/selection-above-text-fragment.txt

focusNode: text
focusOffset: 0
selected: "Hello w"
